8 Refactor Model package for Entity.
