+++ DB-Connector/src/main/java/com/pharos/dbconnector/repository/CandidateRepository.java
@@ -1,6 +1,6 @@
 package com.pharos.dbconnector.repository;
 
-import com.pharos.dbconnector.model.Candidate;
+import com.pharos.dbconnector.entity.Candidate;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
