+++ DB-Connector/src/main/java/com/pharos/dbconnector/repository/DocumentRepository.java
@@ -1,6 +1,6 @@
 package com.pharos.dbconnector.repository;
 
-import com.pharos.dbconnector.model.Document;
+import com.pharos.dbconnector.entity.Document;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
