+++ DB-Connector/src/main/java/com/pharos/dbconnector/repository/GenderRepository.java
@@ -1,6 +1,6 @@
 package com.pharos.dbconnector.repository;
 
-import com.pharos.dbconnector.model.Gender;
+import com.pharos.dbconnector.entity.Gender;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
