+++ DB-Connector/src/main/java/com/pharos/dbconnector/repository/NationalityRepository.java
@@ -1,6 +1,6 @@
 package com.pharos.dbconnector.repository;
 
-import com.pharos.dbconnector.model.Nationality;
+import com.pharos.dbconnector.entity.Nationality;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
