+++ DB-Connector/src/main/java/com/pharos/dbconnector/service/CandidateService.java
@@ -1,6 +1,6 @@
 package com.pharos.dbconnector.service;
 
-import com.pharos.dbconnector.model.Candidate;
+import com.pharos.dbconnector.entity.Candidate;
 
 public interface CandidateService {
     public Candidate createCandidate();
