+++ DB-Connector/src/main/java/com/pharos/dbconnector/service/DocumentService.java
@@ -1,6 +1,6 @@
 package com.pharos.dbconnector.service;
 
-import com.pharos.dbconnector.model.Document;
+import com.pharos.dbconnector.entity.Document;
 
 public interface DocumentService {
